Implemented ApiServiceBase to handle REST API requests Integrated ExchangeRatesApi data service Implemented TimeSeriesApiService to consume timeseries data from external api Added unit test for TimeSeriesApiService Updated Rate and BestRate models Implemented RateService (WIP)

+++ Broker.Application/Models/RateModel.cs
@@ -6,11 +6,11 @@ public class RateModel : ModelBase
 {
     public DateTime Date { get; set; }
 
-    public double Rub { get; set; }
+    public decimal Rub { get; set; }
 
-    public double Eur { get; set; }
+    public decimal Eur { get; set; }
 
-    public double Gbp { get; set; }
+    public decimal Gbp { get; set; }
 
-    public double Jpy { get; set; }
+    public decimal Jpy { get; set; }
 }
